download-project-gutenberg/requirements.txt
google-cloud-storage
bs4
